Make crypt.awk work with other awk(1) variants.
Change use of the coprocess pipe operator, a GNU extension, to a
regular pipe when reading data back from blkid(8). Since we aren't
piping data to blkid(8) and then reading it back, we only need a
regular pipe and getline here anyway.

Without this change, using mawk(1) as the default awk on Void
results in failure when parsing /etc/crypttab and dumps us to
a rescue shell on boot.