vlogger: Fix symlinked ./run handling, make sure tag is initialized.
If vlogger is symlinked to foo/log/run for some service foo,
it's invoked as ./run with $SERVICEDIR/foo/log as the cwd.
So extract the service name from cwd if argv[0] is ./run.
Also tag is initialized to the generic "vlogger" if it is
not given on the command line and servicename couldn't be extracted
from cwd.