services/agetty-generic: use chpst -P instead of setsid
The setsid program has a surprise gotcha of backgrounding a process if
it already happens to be the session leader. This is ok for runit which
never sets the service as a session leader but does not work with other
supervisors which might. chpst provides the same setsid(2) functionality
without also having the gotcha.

Be aware that chpst has issues around gidlist squashing when using
the -u or -U options without explicit passing. This behavior is not
triggered in this case since chpst is being run as root and not being
asked to drop privileges.